Refactor: Adaptations for merging
