+++ src/App.module.css
@@ -1,5 +1,4 @@
 .mainContainer {
-    flex-grow: 1;
     padding-left: 15%;
     padding-right: 15%;
 }
