+++ src/pages/HomePage/data.ts
@@ -29,7 +29,7 @@ export const HelpNeeded = [
         history : "(наполнение 1) Иван появился на свет с довольно редкой патологией сердца – в нем всего один желудочек. Подобная аномалия корректируется (mock up text after this point!!!)",
         toCollect : 108500,
         collected : 8000,
-        image: "src/assets/images/common/test.png"
+        image: "src/assets/images/test.png"
     },
     {
         name : "Андрей Андреев",
@@ -37,7 +37,7 @@ export const HelpNeeded = [
         history : "(наполнение 2) Среди диагнозов есть довольно серьезный, требующий неусыпного наблюдения родителей и врачей – врожденный порок сердца, (mock up text after this point!!!)",
         toCollect : 70000,
         collected : 20000,
-        image: "src/assets/images/common/test.png"
+        image: "src/assets/images/test.png"
     },
     {
         name : "Василий Васильев",
@@ -45,7 +45,7 @@ export const HelpNeeded = [
         history : "(наполнение 3) У мальчика врожденный порок сердца, осложненный легочной гипертензией. В первый год жизни он перенес сложную операцию (mock up text after this point!!!)",
         toCollect : 102000,
         collected : 1000,
-        image: "src/assets/images/common/test.png"
+        image: "src/assets/images/test.png"
     }
 ]
 
@@ -104,16 +104,16 @@ export const News = [
     {
         date : "27.02.2025",
         headline : "Мы выиграли грант!",
-        image: "src/assets/images/common/test.png"
+        image: "src/assets/images/test.png"
     },
     {
         date : "31.01.2025",
         headline : "Алина нуждается в дорогостоящем препарате",
-        image: "src/assets/images/common/test.png"
+        image: "src/assets/images/test.png"
     },
     {
         date : "15.01.2025",
         headline : "Настю ждут в Москве!",
-        image: "src/assets/images/common/test.png"
+        image: "src/assets/images/test.png"
     }
 ]
