+++ src/сomponents/CarouselHome/CarouselHome.tsx
